fix emqx_ws_connection: prevent crashes on pong
Websocket protocol defines its own ping/pong keep alive mechanism in the
spec. Once connected, either end may send a ping, and must receive a
pong. Cowboy responds to pings with pongs automatically, but if a pong
is received, it does nothing.

A pong should only be received in response to a ping, and it isn't clear
where the pings are originating, but failing to handle the pongs causes
this:

```
2019-02-08 08:15:24.201 [error] mqttjs_526dcabe@10.107.155.142:37778 crasher:
    initial call: cowboy_clear:connection_process/5
    pid: <0.36.4>
    registered_name: []
    exception error: no function clause matching
                     emqx_ws_connection:websocket_handle(pong,
                                                         {state,undefined,
                                                          undefined,
            %% <snip>: trimmed error message
```

This commit adds handlers for ping and pong messages to prevent crashes.
Spec requires a pong to be sent in response to ping, cowboy handles this
for us so we don't do anything.