fix(ekka): Upgrade to 0.8.0 to allow remsh
Before ekka 0.8.0, when epmd is not used, ekka_dist would calculate
a port number from node name. This does not work for remote consoles
because a remsh node name is like remsh81random0-<target-node>
i.e. the calculated port number is always the same as the target node.

The fix in 0.8.0 allows remsh prefixed nodes to use a range of
ports to listen.