perf(emqx_trie): reduce remote lock overheads in cluster
The network roundtrip time of taking multiple various locks in emqx_trie
transaction kills the performance of handling a subscribe message.

Take emqx_trie whole table locks to reduce the remote lock overheads in a
cluster setup when there are more than 4 paths in a route.

note 1,

measured on a two nodes on the same host:

without patch: 90 out of 10k clients are able to connect and subscribe,
               others are all timeout.
               (message rate: 100/s, sub timeout: 30s)

with patch:    All 10k clients are able to connect and subscribe.
               (message rate: 100/s, sub timeout: 30s)

note 2, number 4 in this case is an arbitrary number.