feat(gcp_pubsub): generate jwt tokens on demand without workers (rv4.4)
https://emqx.atlassian.net/browse/EMQX-8683

Rather than relying on a JWT worker to produce and refresh tokens, we
could just produce them on demand.  That can generate a bit of extra
work (as multiple processes might realize it’s time to refresh the JWT
and do so), but that shouldn’t be much.  In return, we avoid any
possibility of not having a fresh JWT when pushing messages.