test(crl): add test case for revoke-then-refresh (v4.4)
Investigation for https://github.com/emqx/emqx/issues/9732.

Apparently, there is no problem if a CRL is cached without
revocations, a client cert is revoked, and later the CRL is refreshed.
The test case indicates that the same client can at first connect to
the server and later is denied connection because the CRL was
refreshed.

Also, we take the opportunity in this PR to refresh the test
certificate and CRL files which would expire in June 2023.