revert: Revert "refactor: use mneisa:subscribe/1 for rule operations"
This reverts commit 64309a193be697b4cb62565cda0ad14c040d0dbe.

For relup from old versions, the emqx_rule_registry has not subscribes
the mnesia events, so we cannot update caches only when mnesia events
received.

For rolling upgrade the nodes one by one, some of the nodes has not
subscribes the mnesia events, so we cannot move the `load_hooks_for_rule`
into the mnesia events handling.

In a word, it is too complicated to use mnesia events, we use RPC instead.